+++ plugins/GSdx/GSRendererHW.h
@@ -138,7 +138,7 @@ private:
 
 	#pragma endregion
 
-	int Interpolate_UV(float alpha, int t0, int t1);
+	uint16 Interpolate_UV(float alpha, int t0, int t1);
 	float alpha0(int L, int X0, int X1);
 	float alpha1(int L, int X0, int X1);
